requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cloud-sql-python-connector==0.4.0
colorama==0.4.4
commonmark==0.9.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
gspread==4.0.1
httplib2==0.19.1
idna==3.2
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.4.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
# psycopg2==2.9.1
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pygsheets==2.0.5
pyparsing==2.4.7
pyrsistent==0.18.0
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.9.0
rsa==4.7.2
Send2Trash==1.8.0
six
smmap==4.0.0
streamlit==0.88.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
tzlocal==3.0
uritemplate==3.0.1
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
ImageHash==4.3.1
